Normalize trybuild snapshot target paths
Render staged target-root placeholders with forward slashes before
trybuild compares compile-fail diagnostics. Preserve the platform-neutral
checked-in snapshots while retaining an absolute fallback outside the
workspace.

Cover default, coverage, Windows-style, and outside-workspace roots.
